+++ Enlighter.php
@@ -3,7 +3,7 @@
     Plugin Name: Enlighter - Customizable Syntax Highlighter
     Plugin URI: http://enlighterjs.org
     Description: Enlighter is a free, easy-to-use, syntax highlighting tool with a build-in theme editor.
-    Version: 3.1-BETA2
+    Version: 3.1-BETA3
     Author: Andi Dittrich
     Author URI: http://andidittrich.de
     License: MIT X11-License
@@ -22,8 +22,9 @@
 */
 
 define('ENLIGHTER_INIT', true);
-define('ENLIGHTER_VERSION', '3.1-BETA2');
+define('ENLIGHTER_VERSION', '3.1-BETA3');
 define('ENLIGHTER_PLUGIN_PATH', dirname(__FILE__));
+define('ENLIGHTER_PLUGIN_URL', plugins_url('/enlighter/'));
 
 // check php version
 if (version_compare(phpversion(), '5.3', '>=')){
@@ -34,6 +35,7 @@ if (version_compare(phpversion(), '5.3', '>=')){
     require_once(ENLIGHTER_PLUGIN_PATH.'/class/LegacyShortcodeHandler.php');
     require_once(ENLIGHTER_PLUGIN_PATH.'/class/LowlLevelShortcodeProcessor.php');
     require_once(ENLIGHTER_PLUGIN_PATH.'/class/ResourceLoader.php');
+    require_once(ENLIGHTER_PLUGIN_PATH.'/class/ResourceManager.php');
     require_once(ENLIGHTER_PLUGIN_PATH.'/class/SimpleTemplate.php');
     require_once(ENLIGHTER_PLUGIN_PATH.'/class/CacheManager.php');
     require_once(ENLIGHTER_PLUGIN_PATH.'/class/ThemeGenerator.php');