+++ class/LegacyShortcodeHandler.php
@@ -129,7 +129,7 @@ class LegacyShortcodeHandler{
             
             // tab-name available ?
             if ($shortcodeAttributes['tab']){
-                $htmlAttributes['data-enlighter-tab'] = trim($shortcodeAttributes['tab']);
+                $htmlAttributes['data-enlighter-title'] = trim($shortcodeAttributes['tab']);
             }
             
             // codegroup active ?